requirements.txt
flask
openai
python-pptx
requests
flask-cors
gunicorn
